requirements.txt
requests==2.32.5
jinja2==3.1.6
pyyaml==6.0.3
beautifulsoup4==4.14.3
